feat:所有support

+++ myframework-cloud-framework/myframework-module-common/src/main/java/com/kuma/boot/common/support/generator/EmailAddressGenerator.java
@@ -1,16 +1,39 @@
 /*
- * Decompiled with CFR 0.152.
- * 
- * Could not load the following classes:
- *  org.apache.commons.lang3.RandomStringUtils
+ * Copyright (c) 2020-2030, kuma (2569277704@qq.com & https://blog.kumacloud.top/).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
+
 package com.kuma.boot.common.support.generator;
 
 import com.kuma.boot.common.support.generator.base.GenericGenerator;
 import org.apache.commons.lang3.RandomStringUtils;
 
-public class EmailAddressGenerator
-extends GenericGenerator {
+/*
+ * EmailAddressGenerator
+ * @author kuma
+ * @version 2023.12
+ * @since 2023-12-18 15:18:43
+ */
+/**
+ * EmailAddressGenerator
+ *
+ * @author kuma
+ * @version 2026.01
+ * @since 2025-12-17 10:30:45
+ */
+public class EmailAddressGenerator extends GenericGenerator {
+
     private static final GenericGenerator INSTANCE = new EmailAddressGenerator();
 
     private EmailAddressGenerator() {
@@ -22,8 +45,13 @@ extends GenericGenerator {
 
     @Override
     public String generate() {
-        String result = RandomStringUtils.randomAlphanumeric((int)10) + "@" + RandomStringUtils.randomAlphanumeric((int)5) + "." + RandomStringUtils.randomAlphanumeric((int)3);
+        String result =
+                RandomStringUtils.randomAlphanumeric(10)
+                        + "@"
+                        + RandomStringUtils.randomAlphanumeric(5)
+                        + "."
+                        + RandomStringUtils.randomAlphanumeric(3);
+
         return result.toLowerCase();
     }
 }
-
